pyproject.toml
[tool.poetry]
name = "tv-tracker"
version = "0.1.0"
description = ""
authors = ["Claire Ramming <ceramming@gmail.com>"]
readme = "README.md"
packages = [
  { include = "backend" }
]

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.poetry.dependencies]
python = "^3.10.15"

# Django and service related dependencies
django = "^5.1.5"
djangorestframework = "^3.15.2"
authlib="^1.0"
python-dotenv="^0.19"
requests="^2.27"
dj-database-url="^2.3.0"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
